backend/pyproject.toml
[project]
name = "tagline-backend"
version = "0.1.0"
description = "Tagline Backend"
authors = [
  { name = "Jeffery Harrell" },
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
license = { text = "MIT" }
dependencies = [
    "pydantic-settings>=2.0.0,<3.0.0",
    "alembic>=1.13.0,<2.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "dropbox>=11.36.2",
    "pillow>=10.0.0",
    "pillow-heif>=0.13.0",
    "cachetools>=5.0.0",
    "rq>=1.16.1",
    "rq-dashboard",
    "uvicorn[standard]>=0.29.0,<1.0.0",
    "fastapi>=0.110.0,<1.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "stytch>=13.5.0",
    "pydantic[email]>=2.11.4",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end",
]
addopts = "-ra"
pythonpath = "."

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.12"
reportIncompatibleVariableOverride = false
include = ["app", "tests"]
exclude = ["**/node_modules", "**/__pycache__", "venv"]
reportMissingImports = true
reportUnusedVariable = "warning"
executionEnvironments = [
  { root = "." }
]

[tool.isort]
profile = "black"
known_first_party = "app,tests"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
use_parentheses = true

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[dependency-groups]
dev = [
    "alembic>=1.13.0,<2.0.0",
    "black>=24.3.0,<25.0.0",
    "cachetools>=5.0.0",
    "docker>=7.0.0,<8.0.0",
    "dropbox>=11.36.2",
    "fastapi>=0.110.0,<1.0.0",
    "httpx>=0.27.0,<1.0.0",
    "isort>=6.0.1,<7.0.0",
    "pillow>=10.0.0",
    "pillow-heif>=0.13.0",
    "pre-commit>=3.7.0,<4.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "pyright==1.1.400",
    "pytest>=8.1.0,<9.0.0",
    "pytest-asyncio>=0.26.0,<0.27.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "pytest-order>=1.2.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "requests>=2.31.0,<3.0.0",
    "requests-mock>=1.12.1",
    "rq>=1.16.1",
    "rq-dashboard>=0.8.2.2",
    "ruff>=0.3.0,<1.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "uvicorn[standard]>=0.29.0,<1.0.0",
]
